feat(vix): move the polished editor to userland (vix.elf)
The `vix` command was an in-kernel builtin that ran inside the shell
task, so it never appeared in maktop and ran in ring 0.  The userland
vix.elf existed but was a simpler, older port (no gutter, no wrap,
hardcoded 80x25).

Bring vix.elf to feature parity with the in-kernel editor and make it
the `vix` command:
- Vim-style 5-col line-number gutter, word wrap with '+' continuation,
  '~' past-EOF rows, resolution-agnostic geometry via SYS_TERM_SIZE.
- Richer status bar (Saved / unsaved-warning / Ln X/Y), Ctrl+S save,
  Ctrl+Q quit (double-press when dirty).
- Flashing block caret via a new SYS_CARET_STYLE(218) syscall (sets the
  VESA caret style, returns previous; no-op in VGA-text mode).

Remove the `vix` shell builtin so it resolves via PATH to vix.elf,
running as its own ring-3 task (shows in maktop with pid + memory).
Delete the now-dead in-kernel editor (proc/vix.c, kernel/vix.h) and its
make.config entry.  Docs updated.
