+++ CLAUDE.md
@@ -176,6 +176,7 @@ Authoritative table in `src/kernel/include/kernel/syscall.h`. Selected entries:
 | 158 | SYS_YIELD        | - |
 | 200 | SYS_GETKEY       | raw single-char keyboard read |
 | 201–204 | SYS_PUTCH_AT / SET_CURSOR / TTY_CLEAR / TERM_SIZE | direct TTY ops for full-screen apps (vix) |
+| 218 | SYS_CARET_STYLE | set VESA caret style (0=line, 2=flashing block); returns previous.  No-op in VGA-text mode.  Used by vix.elf |
 | 205 | SYS_WRITE_FILE   | path, buf, len |
 | 206 | SYS_LS_DIR       | path, buf, bufsz |
 | 207 | SYS_DISK_INFO    | buf, bufsz |
@@ -246,7 +247,7 @@ Freestanding ELF binaries built with the cross-compiler. Link against `crt0.S` +
 | `clock.elf` | Fullscreen wall-clock display (CMOS RTC via `/proc/rtc`).  For scripted / one-line use see the `datetime`/`date`/`time` shell builtins instead. |
 | `diskinfo.elf` | partition table + FAT32 BPB dump via `SYS_DISK_INFO` |
 | `fdisk.elf` | MBR partition editor — opens a `/dev` block device, edits the four primary entries (`p`/`n`/`d`/`t`/`a`/`w`/`q`).  `fdisk /dev/hda` (defaults to `/dev/hda`).  Writes the 512-byte MBR back through the block-device fd; devfs's read-modify-write preserves the bootstrap code |
-| `vix.elf` | pane-aware vi-style text editor; uses `SYS_PUTCH_AT` / `SYS_SET_CURSOR` / `SYS_TERM_SIZE` |
+| `vix.elf` | vi-style text editor (the `vix` command — runs as its own ring-3 task, shows in maktop).  Vim-style line-number gutter, word wrap with `+` continuation markers, `~` past-EOF rows, flashing block caret (`SYS_CARET_STYLE`), resolution-agnostic via `SYS_TERM_SIZE`; uses `SYS_PUTCH_AT` / `SYS_SET_CURSOR`.  Ctrl+S save, Ctrl+Q quit (double-press when dirty).  Replaced the former in-kernel `vix` builtin (`proc/vix.c`, removed) |
 | `kbtester.elf` | keyboard diagnostic — logs every event (scancode/keycode/sentinel/modifier) to serial via `SYS_WRITE_SERIAL` |
 | `help.elf` | replaced by `lsman` / `man <cmd>` shell builtins; kept for compatibility |
 
@@ -310,11 +311,11 @@ subsystems:
 
 - **Display**: VESA framebuffer (Bochs VBE, defaults to 720p), VGA text fallback (80×50). Pane abstraction (`vesa_pane_t`) used by VIX. Per-TTY logical character grid (`vt_buf_t` in `display/vt.c`) backs every shell — writes go to the grid first; the framebuffer is only painted when that TTY is focused. After any "fullscreen" shell command returns (vix, install, any ELF launched via `exec` or PATH), `shell_dispatch` calls `shell_restore_screen()` which repaints the focused VT's grid to the FB — so post-exit screen is never blank.
 - **Multi-TTY**: 4 shell tasks (`shell0`–`shell3`). `vtty.c` routes keyboard input via `task_t.tty` (authoritative) and tracks the focused slot. `vtty_switch()` defers the framebuffer repaint out of IRQ context to `vtty_drain_pending()`, which runs from the destination shell's `keyboard_getchar` poll loop. A tmux-style status bar lives in the reserved bottom row showing `Makar  VT0  VT1  VT2  VT3  ...  Alt+F1-F4` with the active slot highlighted.
-- **VIX**: Pane-aware text editor. Derives column/row counts from the active `vesa_pane_t` at runtime - works correctly at any VESA resolution. Modelled on ELKS/FUZIX vi: lightweight, stable, no heap after startup.
+- **VIX**: vi-style text editor, now a **userland** ELF (`vix.elf`, the `vix` command) running as its own ring-3 task — so it appears in maktop with its own pid + memory.  Vim-style line-number gutter, word wrap, `~` EOF rows, and a flashing block caret (`SYS_CARET_STYLE`); derives geometry from `SYS_TERM_SIZE` so it works at any resolution. Modelled on ELKS/FUZIX vi. The earlier in-kernel implementation (`proc/vix.c`) was removed once the userland port reached feature parity.
 - **Storage**: FAT32 (HDD/USB) + ISO 9660 (CD-ROM) via IDE PIO. VFS layer with CWD, auto-mount. Full read/write/delete/rename support on FAT32. Disk filesystems live under `/mnt` (Linux convention): the FAT32 volume at `/mnt/hd` (the default OS drive; mountpoint configurable via `mount`) and the CD-ROM at `/mnt/cdrom`. Bare `/hd` and `/cdrom` remain transitional aliases. Synthetic `/proc` mount exposes `cpuinfo`, `meminfo`, `tasks`, `uname`, `rtc` as read-only files generated on demand.
 - **Block devices (`/dev`)**: Synthetic devfs mount exposing raw IDE storage as byte-addressed nodes: `/dev/hda`, `/dev/hdb`… (whole ATA disks), `/dev/hdaN` (their MBR/GPT partitions, as offset windows), `/dev/cdrom` (ATAPI, read-only). Reads/writes translate to native sector I/O (512 B ATA / 2048 B ATAPI) with internal read-modify-write so callers can transfer arbitrary offset/length. Opening a `/dev` node binds the fd as `FD_KIND_BLOCKDEV` (no eager buffer); `SYS_READ`/`SYS_WRITE`/`SYS_LSEEK` route through `devfs_pread`/`devfs_pwrite`. Backs `fdisk.elf` and `mount /dev/hdaN /mnt/<name>`.
 - **Tasking**: Round-robin scheduler with timer-driven preemption (PIT 100 Hz, `SCHED_QUANTUM = 4` ticks → 40 ms slice). Per-task `pid`, `cwd`, `tty`, signal bitmasks, and real per-task fd table (`fd_table_t` in `kernel/fd.h`, 16 slots, fds 0/1/2 pre-bound to stdin/stdout/stderr). User PD reaped on task exit, fd table reaped on slot reuse. Background ktest harness runs before the shell prompt appears.
-- **Userspace**: Ring-3 protected mode via `iret`. ELF loader (`elf_exec`) with argc/argv. Syscalls: `SYS_EXIT`, `SYS_READ`, `SYS_WRITE` (fd 1 = VGA, fd 2 = VGA + COM1 serial), `SYS_OPEN`, `SYS_CLOSE`, `SYS_LSEEK`, `SYS_BRK`, `SYS_DEBUG`, `SYS_YIELD`, plus Makar extensions (200–215 - terminal/file ops + `SYS_WRITE_SERIAL` + `SYS_GETCWD`). Apps: `calc.elf`, `hello.elf`, `makbox.elf` (ls/cat/cp/mv/rm/rmdir/echo/pwd multicall), `vix.elf`, `diskinfo.elf`, `fdisk.elf`, `clock.elf`, `maktop.elf`, `kbtester.elf`.
+- **Userspace**: Ring-3 protected mode via `iret`. ELF loader (`elf_exec`) with argc/argv. Syscalls: `SYS_EXIT`, `SYS_READ`, `SYS_WRITE` (fd 1 = VGA, fd 2 = VGA + COM1 serial), `SYS_OPEN`, `SYS_CLOSE`, `SYS_LSEEK`, `SYS_BRK`, `SYS_DEBUG`, `SYS_YIELD`, plus Makar extensions (200–218 - terminal/file ops + `SYS_WRITE_SERIAL` + `SYS_GETCWD` + `SYS_CARET_STYLE`). Apps: `calc.elf`, `hello.elf`, `makbox.elf` (ls/cat/cp/mv/rm/rmdir/echo/pwd multicall), `vix.elf`, `diskinfo.elf`, `fdisk.elf`, `clock.elf`, `maktop.elf`, `kbtester.elf`.
 - **Shell**: Inline editing, history, tab completion, Ctrl+C sigint. `lsman` / `man <cmd>` replace `help`. Built-in file ops: `rm`, `rmdir`, `mv`. `uptime` shows humanised h/m/s. `cat /proc/<entry>` for system introspection. `PATH` is a settable shell variable (default `/mnt/cdrom/apps:/mnt/hd/apps`) consulted by command dispatch + tab completion. `mount /dev/hdaN /mnt/<name>` mounts a FAT32 partition at a chosen mountpoint (legacy `mount <drv> <part#>` → `/mnt/hd`); `umount [/mnt/<name>]` flushes + unmounts the FAT32 volume, `umount /mnt/cdrom` ejects the CD. `shutdown`/`reboot` flush + unmount first.
 - **GRUB**: Two-entry menu (Makar OS + Next available device), 5-second timeout.
