+++ src/kernel/arch/i386/make.config
@@ -41,7 +41,6 @@ $(ARCHDIR)/proc/elf.o \
 $(ARCHDIR)/proc/usertest.o \
 $(ARCHDIR)/proc/ktest.o \
 $(ARCHDIR)/proc/installer.o \
-$(ARCHDIR)/proc/vix.o \
 $(ARCHDIR)/proc/vtty.o \
 $(ARCHDIR)/shell/shell.o \
 $(ARCHDIR)/shell/shell_glob.o \
